Feat: #26 implement Commit status model with its tests
