+++ src/main/java/app/ciserver/CommitStatusModel.java
@@ -1,6 +1,6 @@
 package app.ciserver;
 
 public class CommitStatusModel {
-    public record CommitStatus(String state, String description, String context) {
-    }
+	public record CommitStatus(String state, String description, String context) {
+	}
 }
